fix: resolve code quality nuances (unwrap safety, TODOs, schema split)
- Replace hot-path unwrap() calls with safe alternatives:
  - agent/loop_.rs: use if-let for JSON object mutation instead of unwrap()
  - agent/loop_.rs: use unwrap_or_else(|e| e.into_inner()) for mutex locks
    to recover from poison instead of panicking
  - providers/gemini.rs: replace auth_service unwrap() with ok_or_else()
    matching the pattern used elsewhere in the same file

- Resolve or clarify TODO/FIXME comments:
  - security/pairing.rs: replace misleading TODO with design rationale
    (parking_lot + spawn_blocking is the correct pattern for CPU-bound
    sync work); replace expect() on JoinHandle with match for graceful
    failure handling
  - plugins/wasm_tool.rs: replace TODO with explicit fail-fast error
    message; remove unused args parameter
  - plugins/wasm_channel.rs: replace silent TODO stubs with bail!() so
    callers get clear errors instead of silent no-ops
  - channels/transcription.rs: clarify blocked-on-reqwest comment
  - firmware/esp32/src/main.rs: replace dummy return value with explicit
    bail!() so gpio_read fails visibly instead of returning misleading 0

- Split config/schema.rs (13,693 lines) into directory module:
  - Convert schema.rs to schema/mod.rs (9,265 lines)
  - Extract test module to schema/tests.rs (4,427 lines)
  - Make default_true/default_false pub(crate) for submodule access
  - All 235 schema tests pass, zero clippy warnings
